+++ src/main/java/com/dku/council/domain/post/model/dto/PostFileDto.java
@@ -1,7 +1,7 @@
 package com.dku.council.domain.post.model.dto;
 
 import com.dku.council.domain.post.model.entity.PostFile;
-import com.dku.council.infra.nhn.service.ObjectUploadContext;
+import com.dku.council.infra.nhn.s3.service.ObjectUploadContext;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import org.springframework.http.MediaType;
@@ -19,20 +19,15 @@ public class PostFileDto {
     @Schema(description = "파일 url", example = "http://1.2.3.4/1ddee68d-6afb-48d0-9cb6-04a8d8fea4ae.png")
     private final String url;
 
-    @Schema(description = "썸네일 url (없으면 기본 이미지)", example = "http://1.2.3.4/thumb-1ddee68d-6afb-48d0-9cb6-04a8d8fea4ae.png")
-    private final String thumbnailUrl;
-
-    @Schema(description = "원본파일 이름", example = "my_image.png")
+    @Schema(description = "원본파일 이름", example = "my_text.txt")
     private final String originalName;
 
-    @Schema(description = "파일 타입", example = "image/jpeg")
+    @Schema(description = "파일 타입", example = "text/plain")
     private final String mimeType;
-
 
     public PostFileDto(ObjectUploadContext context, PostFile file) {
         this.id = file.getId();
-        this.url = context.getObjectUrl(file.getFileId());
-        this.thumbnailUrl = context.getThumbnailUrl(file.getThumbnailId());
+        this.url = context.getFileUrl(file.getFileId());
         this.originalName = file.getFileName();
 
         String fileMimeType = file.getMimeType();
