--- src/main/java/com/dku/council/infra/nhn/exception/InvalidFileContentTypeException.java
@@ -1,10 +1,0 @@
-package com.dku.council.infra.nhn.exception;
-
-import com.dku.council.global.error.exception.LocalizedMessageException;
-import org.springframework.http.HttpStatus;
-
-public class InvalidFileContentTypeException extends LocalizedMessageException {
-    public InvalidFileContentTypeException(Throwable e) {
-        super(e, HttpStatus.BAD_REQUEST, "invalid.file-content-type");
-    }
-}
