+++ src/main/java/com/dku/council/infra/nhn/global/exception/CannotSendSMSException.java
@@ -1,0 +1,14 @@
+package com.dku.council.infra.nhn.global.exception;
+
+import com.dku.council.global.error.exception.LocalizedMessageException;
+import org.springframework.http.HttpStatus;
+
+public class CannotSendSMSException extends LocalizedMessageException {
+    public CannotSendSMSException() {
+        super(HttpStatus.BAD_REQUEST, "failed.send-sms");
+    }
+
+    public CannotSendSMSException(Throwable t) {
+        super(t, HttpStatus.BAD_REQUEST, "failed.send-sms");
+    }
+}
