+++ src/main/java/com/dku/council/infra/nhn/global/exception/InvalidFileContentTypeException.java
@@ -1,0 +1,10 @@
+package com.dku.council.infra.nhn.global.exception;
+
+import com.dku.council.global.error.exception.LocalizedMessageException;
+import org.springframework.http.HttpStatus;
+
+public class InvalidFileContentTypeException extends LocalizedMessageException {
+    public InvalidFileContentTypeException(Throwable e) {
+        super(e, HttpStatus.BAD_REQUEST, "invalid.file-content-type");
+    }
+}
